Adjust left-right pane ratio, apply min-height to body, set section width
The left pane got completely squashed in mobile device. Hence, adjusted the
width ratio of left and right panes (header and main elements.)

Applied a minimum height of 98vh to body in order to spread things out for a
neater look.

Set section width as the text content went all the way to the edge of screen on
desktop, looking rather ugly.
The section will take up a width of 65ch on wide screens while 100% of the width
on narrow screens. This helps maintain a responsive design.