Create section-width variable, apply basic styling to form elements
The background color of form section was distinct and since it didn't
take up all the width of its container, it stood out in an undesirable
way.
This pressed the need to create a '--section-width' variable. Doing so
allowed me to set form container width to max while setting the actual
width of form the same as other sections.
This way, the background-color seemed more natural in its place and the
form content didn't splurge all over the screen and remained in alignment
with content of other sections.

Applied basic styling to button, label, and input elements.