Add pattern for phone number and password validation
Set phone number pattern so that it only contains valid phone
numbers with 10 digits or 11 digits in case the number starts
with 0.
More advanced validation will be added in future.

Set password pattern so that it is only valid if it contains a
lower-case letter, an upper-case letter, a digit, a symbol, and
is at least 8 characters long.