Turn all required and empty inputs inaccessible to :invalid pseudo-class
Initially, all the inputs that were empty and had the 'required'
attribute set on them were being targeted by :invalid pseudo-class.
This led to the :invalid styling being forcefully applied on all
empty inputs.

Therefore a class of 'empty' is applied to all empty inputs and :invalid
styling is done to only those inputs that do not have that class (which
is toggled with the help of event listener.)