Construct login panel and apply appropriate styling
Panel has been constructed with position set to absolute so that it
appears over the DOM when called.
The visibility of panel has been set to hidden until a close button
is created.

The panel copies the header background image and heading.

The panel also copies the styling of form elements from the form
to create account.

A box-shadow with maximum spread has been added so that the background
fades away when login panel is called.