Add login panel open and close feature
Added event listener to "log-in" text on main page to open the login
panel by changing its visibility to "visible".

Added a close button and placed it on the top-right corner of login
panel by creating an empty div at the top to serve as its relative
position.
Close button closes the login panel by changing the visibility property
to "hidden" upon click.